requirements.txt
flask==2.2.2
loguru
pyyaml
